Add files via upload
PIC Thermometer and Thermostat

The circuit is designed for a very specific operation: For the etching bath heater. I use this circuit for my homemade etching bath. A thermostat is a device that measures the temperature (in this case the temperature of the etching liquid), and according to a temperature window, will turn on or off the heater. This circuit has a fixed temperature window between 37 and 41oC. This can be changed of course very easy. I am preparing an etching bath and got an aquarium heater. Unfortunately, the heater's thermostat can be set up to 32oC. Higher temperatures are not proper for fish aquariums (unless you think of a fish soup). Therefore, i had to design my own thermostat.